+++ src/senaite/core/registry/schema.py
@@ -22,8 +22,8 @@ from bika.lims import senaiteMessageFactory as _
 from senaite.core.config.worksheet import DEFAULT_WORKSHEET_LAYOUT
 from plone.autoform import directives
 from plone.supermodel import model
+from senaite.core.config.registry import SKIP_ANALYSES_STATES_ON_COPY
 from z3c.form.browser.checkbox import CheckBoxFieldWidget
-
 from zope import schema
 
 
@@ -460,9 +460,42 @@ class ISampleRegistry(ISenaiteRegistry):
             default=u"Samples"
         ),
         fields=[
+            "sample_add_form_skip_partition_analyses",
+            "sample_add_form_skip_analyses_in_states",
             "sample_add_form_allow_multi_paste",
             "trigger_events_on_sample_creation",
         ],
+    )
+
+    sample_add_form_skip_partition_analyses = schema.Bool(
+        title=_(
+            u"label_registry_sample_add_skip_partition_analyses",
+            default=u"Skip partition analyses on copy"
+        ),
+        description=_(
+            u"description_registry_sample_add_skip_partition_analyses",
+            default=u"When enabled, analyses from partitions will be excluded "
+                    u"when copying a sample. Only analyses that directly belong "
+                    u"to the source sample will be copied to the new sample."
+        ),
+        default=False,
+        required=False,
+    )
+
+    sample_add_form_skip_analyses_in_states = schema.List(
+        title=_(
+            u"label_registry_sample_add_skip_analyses_in_states",
+            default=u"Skip analyses workflow states on copy"
+        ),
+        description=_(
+            u"description_registry_sample_add_skip_analyses_in_states",
+            default=u"Add all analyses workflow states that should be "
+                    u"skipped when copying analyses to a new sample in the "
+                    u"sample add form."
+        ),
+        value_type=schema.ASCIILine(),
+        required=False,
+        default=SKIP_ANALYSES_STATES_ON_COPY,
     )
 
     sample_add_form_allow_multi_paste = schema.List(