+++ src/senaite/core/schema/interfaces.py
@@ -26,6 +26,7 @@ from zope.schema.interfaces import IField
 from zope.schema.interfaces import IInt
 from zope.schema.interfaces import IList
 from zope.schema.interfaces import INativeString
+from zope.schema.interfaces import ITextLine
 from zope.schema.interfaces import ITimedelta
 
 
@@ -73,8 +74,12 @@ class IRichTextField(IRichText):
     """Senaite rich text field
     """
 
+class IEmailField(ITextLine):
+    """Input type "email" widget
+    """
 
-class IPhoneField(INativeString):
+
+class IPhoneField(ITextLine):
     """Input type "phone" widget
     """